remove index on assignments.integration_id
It turns out the integration_id is not unique, and in addition,
is not used for lookup by SIS or pro serv.

Fixes SIS-287

Reviewed-on: https://gerrit.instructure.com/35339
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>

+++ db/migrate/20140404162351_add_assignment_post_columns.rb
@@ -5,11 +5,14 @@ class AddAssignmentPostColumns < ActiveRecord::Migration
     add_column :assignments, :post_to_sis, :boolean
     add_column :assignments, :integration_id, :string
 
-    add_index :assignments, :integration_id, unique: true
+    # We used to add an index on integration_id here, but decided not
+    # to add it at all after it'd already been migrated in some envs
   end
 
   def self.down
-    remove_index :assignments, :integration_id
+    if index_exists?(:assignments, :integration_id)
+      remove_index :assignments, :integration_id
+    end
     
     remove_column :assignments, :post_to_sis
     remove_column :assignments, :integration_id
